Swagger doc + some refactoring. #12
