+++ assignment/src/main/java/com/bluedrop/assignment/web/exception/NotFoundException.java
@@ -1,4 +1,4 @@
-package com.bluedrop.assignment.web.controllers;
+package com.bluedrop.assignment.web.exception;
 
 public class NotFoundException extends RuntimeException {
 }
